removed comment on dashboardModal. Moved dashboardModal out of date folder.
